requirements.txt
openai>=1.40.0
requests
python-dotenv
streamlit
pandas
plotly

# Hugging Face
huggingface-hub
transformers
torch

# Optional performance
accelerate
bitsandbytes
